Fixing missing props and rename chatcard file
